feat: consolidate sign-in functionality into the homepage and remove redundant sign-in page
