+++ apps/ned/src/app/page.tsx
@@ -1,97 +1,135 @@
-import Image from 'next/image'
-import Link from 'next/link'
+'use client'
+
+import { useEffect, useState } from 'react'
+
+import { useRouter } from 'next/navigation'
+
+import { useAuthContext } from '@/context/AuthContext'
+import { sendSignInLink, signInWithGoogle } from '@/firebase/auth/signin'
 
 export default function Home() {
+  const { user, loading } = useAuthContext()
+  const router = useRouter()
+  const [email, setEmail] = useState('')
+  const [isLinkSent, setIsLinkSent] = useState(false)
+  const [errorMessage, setErrorMessage] = useState('')
+
+  useEffect(() => {
+    if (!loading && user) {
+      router.push(`${process.env.NEXT_PUBLIC_SANSA_URL}/`)
+    }
+  }, [loading, user, router])
+
+  const handleGoogleLogin = async () => {
+    const { success, error } = await signInWithGoogle()
+
+    if (success) {
+      router.push(`${process.env.NEXT_PUBLIC_SANSA_URL}/`)
+    } else {
+      setErrorMessage(
+        (error as Error)?.message || 'Falha ao fazer login com o Google.'
+      )
+    }
+  }
+
+  const handleForm = async (event: { preventDefault: () => void }) => {
+    event.preventDefault()
+
+    const { success, error } = await sendSignInLink(email)
+
+    if (!success) {
+      setErrorMessage(
+        error === 'auth/email-already-in-use'
+          ? 'Este e-mail já está vinculado a outra forma de login. Por favor, use sua senha ou outro método.'
+          : (error as Error)?.message || 'Falha ao enviar o link de login.'
+      )
+      return
+    }
+
+    setIsLinkSent(true)
+  }
+
+  if (loading || (!loading && user)) {
+    // Tela de carregamento enquanto verifica o estado de autenticação
+    return (
+      <div className="flex items-center justify-center h-screen">
+        <p>Carregando...</p>
+      </div>
+    )
+  }
+
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <div className="z-10 w-full max-w-5xl items-center justify-between font-mono text-sm lg:flex">
-        <div className="fixed left-0 top-0 flex w-full justify-center border-b border-gray-300 bg-gradient-to-b from-zinc-200 pb-6 pt-8 backdrop-blur-2xl dark:border-neutral-800 dark:bg-zinc-800/30 dark:from-inherit lg:static lg:w-auto  lg:rounded-xl lg:border lg:bg-gray-200 lg:p-4 lg:dark:bg-zinc-800/30">
-          <Link href="/signin">
-            <span className="link">Entrar</span>
-          </Link>
+    <div className="flex flex-col items-center justify-center h-screen">
+      <div className="w-full max-w-xs">
+        {!isLinkSent ? (
+          <form
+            onSubmit={handleForm}
+            className="bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4"
+          >
+            <h1 className="text-3xl font-bold mb-6 text-black">Sign In</h1>
+            <div className="mb-4">
+              <label
+                htmlFor="email"
+                className="block text-gray-700 text-sm font-bold mb-2"
+              >
+                Email
+              </label>
+              <input
+                onChange={e => setEmail(e.target.value)}
+                required
+                type="email"
+                name="email"
+                id="email"
+                placeholder="example@mail.com"
+                className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+              />
+            </div>
+            {errorMessage && (
+              <p className="text-red-500 text-sm mb-4">{errorMessage}</p>
+            )}
+            {errorMessage ===
+              'Este e-mail já está vinculado a outra forma de login. Por favor, use sua senha ou outro método.' && (
+              <p className="text-gray-700">
+                Já possui uma conta?{' '}
+                <a href="/login" className="text-blue-500">
+                  Faça login aqui
+                </a>
+                .
+              </p>
+            )}
+            <div className="flex items-center justify-between mt-4">
+              <button
+                type="submit"
+                className="w-full bg-blue-500 text-white font-semibold py-2 rounded"
+              >
+                Enviar Link de Login
+              </button>
+            </div>
+          </form>
+        ) : (
+          <div className="bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4">
+            <h1 className="text-2xl font-bold mb-6 text-black">Link Enviado</h1>
+            <p className="text-gray-700">
+              Verifique sua caixa de entrada e clique no link enviado para
+              concluir o login.
+            </p>
+            <button
+              onClick={() => router.push('/')}
+              className="mt-4 w-full bg-blue-500 text-white font-semibold py-2 rounded"
+            >
+              Voltar para a página inicial
+            </button>
+          </div>
+        )}
+        <div className="mt-4">
+          <button
+            onClick={handleGoogleLogin}
+            className="w-full bg-red-500 text-white font-semibold py-2 rounded"
+          >
+            Login com Google
+          </button>
         </div>
       </div>
-
-      <div className="relative flex place-items-center before:absolute before:h-[300px] before:w-[480px] before:-translate-x-1/2 before:rounded-full before:bg-gradient-radial before:from-white before:to-transparent before:blur-2xl before:content-[''] after:absolute after:-z-20 after:h-[180px] after:w-[240px] after:translate-x-1/3 after:bg-gradient-conic after:from-sky-200 after:via-blue-200 after:blur-2xl after:content-[''] before:dark:bg-gradient-to-br before:dark:from-transparent before:dark:to-blue-700 before:dark:opacity-10 after:dark:from-sky-900 after:dark:via-[#0141ff] after:dark:opacity-40 before:lg:h-[360px]">
-        <Image
-          className="relative dark:drop-shadow-[0_0_0.3rem_#ffffff70] dark:invert"
-          src="/next.svg"
-          alt="Next.js Logo"
-          width={180}
-          height={37}
-          priority
-        />
-      </div>
-
-      <div className="mb-32 grid text-center lg:mb-0 lg:grid-cols-4 lg:text-left">
-        <a
-          href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <h2 className={`mb-3 text-2xl font-semibold`}>
-            Docs{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
-            Find in-depth information about Next.js features and API.
-          </p>
-        </a>
-
-        <a
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800 hover:dark:bg-opacity-30"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <h2 className={`mb-3 text-2xl font-semibold`}>
-            Learn{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
-            Learn about Next.js in an interactive course with&nbsp;quizzes!
-          </p>
-        </a>
-
-        <a
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <h2 className={`mb-3 text-2xl font-semibold`}>
-            Templates{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
-            Explore the Next.js 13 playground.
-          </p>
-        </a>
-
-        <a
-          href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          className="group rounded-lg border border-transparent px-5 py-4 transition-colors hover:border-gray-300 hover:bg-gray-100 hover:dark:border-neutral-700 hover:dark:bg-neutral-800/30"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <h2 className={`mb-3 text-2xl font-semibold`}>
-            Deploy{' '}
-            <span className="inline-block transition-transform group-hover:translate-x-1 motion-reduce:transform-none">
-              -&gt;
-            </span>
-          </h2>
-          <p className={`m-0 max-w-[30ch] text-sm opacity-50`}>
-            Instantly deploy your Next.js site to a shareable URL with Vercel.
-          </p>
-        </a>
-      </div>
-    </main>
+    </div>
   )
 }
